Time slots type changed to long instead of int

+++ src/relations/ChargesAt.java
@@ -1,47 +1,47 @@
 package relations;
 
 public class ChargesAt {
-    private String CarPlate;
-    private int UID;
-    private int TimeStart;
-    private int TimeFinish;
+	private String CarPlate;
+	private int UID;
+	private long TimeStart;
+	private long TimeFinish;
 
-    public ChargesAt(String carPlate, int UID, int timeStart, int timeFinish) {
-        CarPlate = carPlate;
-        this.UID = UID;
-        TimeStart = timeStart;
-        TimeFinish = timeFinish;
-    }
+	public ChargesAt(String carPlate, int UID, long timeStart, long timeFinish) {
+		CarPlate = carPlate;
+		this.UID = UID;
+		TimeStart = timeStart;
+		TimeFinish = timeFinish;
+	}
 
-    public String getCarPlate() {
-        return CarPlate;
-    }
+	public String getCarPlate() {
+		return CarPlate;
+	}
 
-    public void setCarPlate(String carPlate) {
-        CarPlate = carPlate;
-    }
+	public void setCarPlate(String carPlate) {
+		CarPlate = carPlate;
+	}
 
-    public int getUID() {
-        return UID;
-    }
+	public int getUID() {
+		return UID;
+	}
 
-    public void setUID(int UID) {
-        this.UID = UID;
-    }
+	public void setUID(int UID) {
+		this.UID = UID;
+	}
 
-    public int getTimeStart() {
-        return TimeStart;
-    }
+	public long getTimeStart() {
+		return TimeStart;
+	}
 
-    public void setTimeStart(int timeStart) {
-        TimeStart = timeStart;
-    }
+	public void setTimeStart(long timeStart) {
+		TimeStart = timeStart;
+	}
 
-    public int getTimeFinish() {
-        return TimeFinish;
-    }
+	public long getTimeFinish() {
+		return TimeFinish;
+	}
 
-    public void setTimeFinish(int timeFinish) {
-        TimeFinish = timeFinish;
-    }
+	public void setTimeFinish(long timeFinish) {
+		TimeFinish = timeFinish;
+	}
 }
